+++ pkg/providers/appsec/testdata/ase/appsec-import.sh
@@ -1,5 +1,4 @@
 terraform init
-terraform import module.security.akamai_appsec_configuration.config 79947
 terraform import module.security.akamai_appsec_advanced_settings_logging.logging 79947
 terraform import module.security.akamai_appsec_advanced_settings_attack_payload_logging.attack_payload_logging 79947
 terraform import module.security.akamai_appsec_advanced_settings_prefetch.prefetch 79947