+++ pkg/providers/appsec/testdata/ase/modules/security/main.tf
@@ -1,9 +1,5 @@
-resource "akamai_appsec_configuration" "config" {
-  name        = var.name
-  description = var.description
-  contract_id = var.contract_id
-  group_id    = trimprefix(data.akamai_group.group.id, "grp_")
-  host_names  = var.hostnames
+data "akamai_appsec_configuration" "config" {
+  name = var.name
 }
 
 data "akamai_group" "group" {